[ConstraintSystem] Do not truncate variable indices to 16 bits
ConstraintSystem::Entry::Id was uint16_t, while variable indices are
unsigned everywhere else: they come from Value2Index, which is a
DenseMap<Value *, unsigned>, and addVariableRow builds entries from a
size_t index. An index above 65535 was silently truncated on the way
into Entry, and getLastCoefficient took its Id parameter as uint16_t as
well, truncating the index a second time during Fourier-Motzkin
elimination.

The result is a miscompile, not just a missed fold: a fact recorded for
variable 65537 lands on variable 1, so ConstraintElimination folds a
condition on one value using a fact that only constrains an unrelated
one. For a function with 65537 arguments,
Value2Index[%a65536] == 65537 truncates to Value2Index[%a0] == 1, and
"%a65536 <s 5" is enough to fold "%a0 <s 5" to true. The
-constraint-elimination-max-rows cap does not prevent this, as it is
per-system and the argument rows only seed the unsigned system.

Widen Id to unsigned. This is free: Coefficient is an int64_t, so the
struct is padded out to a multiple of its alignment either way, and
sizeof(Entry) is unchanged. The alternative of dropping facts whose
index does not fit would also be sound, but would lose folds for no
benefit. A static_assert documents that Entry is meant to stay small.

The trigger needs more than 65535 variables, and every variable needs a
distinct IR value, so the smallest IR reproducer is a function with
65537 arguments (~840KB) - too large for the test suite. Test the
solver directly instead.

+++ llvm/lib/Analysis/ConstraintSystem.cpp
@@ -92,7 +92,7 @@ bool ConstraintSystem::eliminateUsingFM() {
         // Starts with index 0 and updates every coefficients.
         int64_t UpperV = 0;
         int64_t LowerV = 0;
-        uint16_t CurrentId = std::numeric_limits<uint16_t>::max();
+        unsigned CurrentId = std::numeric_limits<unsigned>::max();
         if (IdxUpper < UpperRow.size()) {
           CurrentId = std::min(UpperRow[IdxUpper].Id, CurrentId);
         }
